fix: gif search and refactor gif modal component to be resuable
