+++ src/pages/projectView/projectMainBody/components/ProjectFundingFormCommentField.tsx
@@ -1,37 +1,25 @@
-import { CloseIcon, SearchIcon } from '@chakra-ui/icons'
+import { CloseIcon } from '@chakra-ui/icons'
 import {
   Box,
   Button,
   HStack,
   HTMLChakraProps,
   Image,
-  Input,
-  InputGroup,
-  InputLeftElement,
-  InputRightElement,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalOverlay,
   Text,
   Tooltip,
   useBoolean,
   useDisclosure,
 } from '@chakra-ui/react'
-import { GiphyFetch } from '@giphy/js-fetch-api'
 import { IGif } from '@giphy/js-types'
-import { Grid } from '@giphy/react-components'
-import { useCallback, useRef, useState } from 'react'
+import { useRef, useState } from 'react'
 
 import { GifIcon } from '../../../../components/icons'
 import { TextArea } from '../../../../components/ui'
-import { VITE_APP_GIPHY_API_KEY } from '../../../../constants'
 import { useAuthContext } from '../../../../context'
 import { useModal } from '../../../../hooks/useModal'
-import { useMobileMode } from '../../../../utils'
 import { LogoutConfirmationModal } from '../../../auth/components/LogoutConfirmationModal'
 import { AvatarElement } from './AvatarElement'
+import { GifModal } from './GifModal'
 
 type Props = HTMLChakraProps<'div'> & {
   comment: string
@@ -39,15 +27,12 @@ type Props = HTMLChakraProps<'div'> & {
   setFormState: any
 }
 
-const giphy = new GiphyFetch(VITE_APP_GIPHY_API_KEY)
-
 export const ProjectFundingFormCommentField = ({
   comment,
   setTarget,
   setFormState,
   ...rest
 }: Props) => {
-  const isMobile = useMobileMode()
   const { isAnonymous, loginOnOpen, user } = useAuthContext()
   const textAreaRef = useRef<HTMLTextAreaElement>(null)
 
@@ -64,18 +49,16 @@ export const ProjectFundingFormCommentField = ({
     onGIFModalOpened()
   }
 
-  const [gifSearch, setGifSearch] = useState('bitcoin')
   const [selectedGIF, setSelectedGIF] = useState<IGif | null>(null)
 
   const [isHoveringOverGIFButton, setIsHoveringOverGIFButton] =
     useBoolean(false)
-  const [focus, setFocus] = useState(true)
 
-  const fetchGifs = useCallback(
-    (offset: number) =>
-      giphy.search(gifSearch, { offset, sort: 'relevant', limit: 9 }),
-    [gifSearch],
-  )
+  const handleSelectGif = (gif: IGif) => {
+    setSelectedGIF(gif)
+    setFormState('media', gif.images.original.webp)
+    onGIFModalClosed()
+  }
 
   return (
     <Box {...rest}>
@@ -158,64 +141,12 @@ export const ProjectFundingFormCommentField = ({
         </Box>
       </HStack>
 
-      {isGIFModalOpen && (
-        <Modal
-          onClose={() => {
-            setGifSearch('bitcoin')
-            onGIFModalClosed()
-          }}
-          isOpen={isGIFModalOpen}
-          isCentered
-        >
-          <ModalOverlay />
+      <GifModal
+        isOpen={isGIFModalOpen}
+        onClose={onGIFModalClosed}
+        onSelect={handleSelectGif}
+      />
 
-          <ModalContent mt={focus && isMobile ? 100 : 0}>
-            <ModalBody p={2}>
-              <InputGroup mb={2}>
-                <InputLeftElement>
-                  <SearchIcon />
-                </InputLeftElement>
-                <Input
-                  onFocus={() => setFocus(true)}
-                  onBlur={() => setFocus(false)}
-                  placeholder="Search"
-                  variant="filled"
-                  focusBorderColor="primary.400"
-                  bg="neutral.400"
-                  onChange={(e) => setGifSearch(e.target.value)}
-                />
-                <InputRightElement mb={2}>
-                  <ModalCloseButton>
-                    <CloseIcon mb={2} />
-                  </ModalCloseButton>
-                </InputRightElement>
-              </InputGroup>
-
-              <Box height="450px" overflow="auto">
-                <Box
-                  display="flex"
-                  justifyContent="center"
-                  alignItems="center"
-                  cursor="pointer"
-                >
-                  <Grid
-                    width={isMobile ? 350 : 400}
-                    columns={3}
-                    fetchGifs={fetchGifs}
-                    noLink={true}
-                    hideAttribution={true}
-                    onGifClick={(gif: any) => {
-                      setSelectedGIF(gif)
-                      setFormState('media', gif.images.original.webp)
-                      onGIFModalClosed()
-                    }}
-                  />
-                </Box>
-              </Box>
-            </ModalBody>
-          </ModalContent>
-        </Modal>
-      )}
       <LogoutConfirmationModal {...logoutConfirmationModal} />
     </Box>
   )
